maint: massive lint and restructuring
