+++ .flake8
@@ -7,4 +7,9 @@ exclude =
 max-line-length = 88
 extend-ignore =
     E203,
+# flake-docstrings
+per-file-ignores =
+    src/dipcoatimage/finitedepth/version.py:D
+    tests/**:D
+    doc/source/conf.py:D
 docstring-convention = numpy
