+++ src/dipcoatimage/finitedepth/util/__init__.py
@@ -1,10 +1,1 @@
-"""
-Utilities
-=========
-
-:mod:`dipcoatimage.finitedepth.util` provides utility functions.
-
-.. automodule:: dipcoatimage.finitedepth.util.parameters
-   :members:
-
-"""
+"""Utility functions."""
